Updated services and tests for doctor and prescription APIs

+++ sql/tables.sql
@@ -33,7 +33,7 @@ create table if not exists address (
 );
 
 create table if not exists pharmacists (
-    pharmacist_id BIGINT primary key,
+    pharmacist_id serial primary key,
     first_name VARCHAR(20) not null,
     last_name VARCHAR(30) not null,
     pharmacist_email VARCHAR(30) not null,
